fallback to okteto.yaml before failing
